+++ sl-map-apis/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+update sl-types dependency to 0.1.1
+
 ## 0.1.0
 
 Initial Release