+++ sl-map-cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+Print PPS HUD config string when generating a map
+
 ## 0.1.0
 
 Initial Release