+++ sl-types/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+add function to `GridRectangleLike` that returns the PPS HUD config string
+for a given grid rectangle
+
 ## 0.1.0
 
 Initial Release